+++ src/lib/constants.js
@@ -15,6 +15,12 @@ const RE_TRAILING_SEMICOLON = /;$/;
 const RE_AMP_ENTITY = /&(#?[0-9a-zA-Z]+;)/g;
 const RE_LEGACY_ENTITIES = /&((?:Iacute|aacute|uacute|plusmn|Otilde|otilde|agrave|Agrave|Yacute|yacute|Oslash|oslash|atilde|Atilde|brvbar|ccedil|Ccedil|Ograve|curren|divide|eacute|Eacute|ograve|Oacute|egrave|Egrave|Ugrave|frac12|frac14|frac34|ugrave|oacute|iacute|Ntilde|ntilde|Uacute|middot|igrave|Igrave|iquest|Aacute|cedil|laquo|micro|iexcl|Icirc|icirc|acirc|Ucirc|Ecirc|ocirc|Ocirc|ecirc|ucirc|Aring|aring|AElig|aelig|acute|pound|raquo|Acirc|times|THORN|szlig|thorn|COPY|auml|ordf|ordm|Uuml|macr|uuml|Auml|ouml|Ouml|para|nbsp|euml|quot|QUOT|Euml|yuml|cent|sect|copy|sup1|sup2|sup3|iuml|Iuml|ETH|shy|reg|not|yen|amp|AMP|REG|uml|eth|deg|gt|GT|LT|lt)(?!;)|(?:#?[0-9a-zA-Z]+;))/g;
 const RE_ESCAPE_LT = /</g;
+// In escapable raw text only the element’s own end tag ends the element, so nothing else
+// that `decodeEntities` decoded to a `<` needs escaping there
+const RE_ESCAPE_LT_RAW_TEXT = /** @type {Record<string, RegExp>} */ ({
+  textarea: /<(?=\/textarea[\s/>])/gi,
+  title: /<(?=\/title[\s/>])/gi
+});
 const RE_ATTR_WS_CHECK = /[ \n\r\t\f]/;
 const RE_ATTR_WS_COLLAPSE = /[ \n\r\t\f]+/g;
 const RE_ATTR_WS_TRIM = /^[ \n\r\t\f]+|[ \n\r\t\f]+$/g;
@@ -172,6 +178,10 @@ const htmlElements = new Set(['a', 'abbr', 'acronym', 'address', 'applet', 'area
 
 const specialContentElements = new Set(['script', 'style']);
 
+// Escapable raw text elements, whose content is text rather than markup
+// https://html.spec.whatwg.org/multipage/syntax.html#elements-2
+const escapableRawTextElements = new Set(['textarea', 'title']);
+
 // Exports
 
 export {
@@ -191,6 +201,7 @@ export {
   RE_AMP_ENTITY,
   RE_LEGACY_ENTITIES,
   RE_ESCAPE_LT,
+  RE_ESCAPE_LT_RAW_TEXT,
   RE_ATTR_WS_CHECK,
   RE_ATTR_WS_COLLAPSE,
   RE_ATTR_WS_TRIM,
@@ -240,5 +251,6 @@ export {
   htmlElements,
 
   // Special content elements
-  specialContentElements
+  specialContentElements,
+  escapableRawTextElements
 };